Move GetBinFWVersion function from excel.py fix excel didn't close issue.
