+++ ReleasePkgLib/__init__.py
@@ -8,12 +8,13 @@ from .Lib import MatchMultipleFolder, ChangeBuildID, RemoveOldFileInDir, Copy_Re
 from .InputStr import InputStr
 from .Zip import UnZip, PrintZipAllInfo, GetZipDateInfo
 from .Ftp import Ftp_multi, Ftp_download, Ftp_download_Test, Ftp_connect, Ftp_get_file_list, Ftp_get_filename, Ftp_callback
-from .Excel import CheckBiosBuildDate, ModifyExcelData, FindOldMEVersion, GetBinaryData, GetMrcVersion, GetMEVersion, GetIshVersion, GetPmcVersion, GetNphyVersion, GetSphyVersion, GetPchcVersion, PrintBiosBuildDate, CheckFileChecksum, PrintBiosBinaryChecksum, CheckMEVersion, ModifyReleaseNote
+from .Excel import CheckBiosBuildDate, ModifyExcelData, FindOldMEVersion, PrintBiosBuildDate, CheckFileChecksum, PrintBiosBinaryChecksum, CheckMEVersion, ModifyReleaseNote
 from .Platform import Platform_Flag
 from .Argparse import argparse_function
 from .Version import Version
 from .Check import CheckPkg, CheckPkg_AMD
 from .Tool_Version_Compare import GetFileInfo, ReadToolVersionTable, SetToolVersionTable, CompareInfo, ChangeVersionInfo, ChangeDataInfo
+from .GetBinFWVersion import GetBinaryData, GetMrcVersion, GetMEVersion, GetIshVersion, GetPmcVersion, GetNphyVersion, GetSphyVersion, GetPchcVersion
 
 
 def main_init():
